Test Containers

+++ src/main/java/ar/com/redlink/services/IGenderService.java
@@ -2,5 +2,5 @@ package ar.com.redlink.services;
 
 import ar.com.redlink.services.impl.entities.GenderEntity;
 
-public interface IGenderService extends ICRUD<GenderEntity, Integer>{
+public interface IGenderService extends ICRUD<GenderEntity, Long>{
 }
