+++ src/main/java/ar/com/redlink/services/IUserTypeIdService.java
@@ -2,6 +2,6 @@ package ar.com.redlink.services;
 
 import ar.com.redlink.services.impl.entities.UserTypeIdEntity;
 
-public interface IUserTypeIdService extends ICRUD<UserTypeIdEntity, Integer>{
+public interface IUserTypeIdService extends ICRUD<UserTypeIdEntity, Long>{
 
 }
